Fix synonym update before index.
